Refactor Adventurer class properties and constructor
Updated the `Adventurer` class to replace the `Toughness` property with an `Armour` property. The constructor now sets default values for `Health`, `Strength`, and initializes `Armour` to 0, while removing the parameters for these properties. The `Wealth` property is initialized to 0, and `Inventory` is set as a new list of `Entity`. The `CurrentRoom` property is also initialized to null.